home_automation_voice_assistant/requirements.txt
paho-mqtt
SpeechRecognition
python3-pyaudio
pyttsx3